Experiment with new sampling method
I thought I could use the coord/val extraction step as an input to the gridding step. This was right, but makes things challenging when you want to crop the grid, then have to resample the coord val parts.

I'm leaving this here as a reminder of how to do it, but want to revert to the method used that is more like in the Source material.