Generate file list once to speed up workers
I thought this was responsible for slow intitial performance of the dataloaders. After learning to profile, and profiling the changes, it offers a small benefit, but is not the main culprit.

I think the culprit is a large memory allocation being pickled for each worker? But I'm going to change to an IterableDataset now anyway.